+++ src/invoices/VersionHistory/VersionHistoryView/VersionHistoryViewInformation/index.js
@@ -1,0 +1,1 @@
+export { VersionHistoryViewInformation } from './VersionHistoryViewInformation';